requirements.txt
Flask==1.1.1
pytest>=4.6.0
optimizely-sdk[security]==3.7.1
